feat: OrderController add

+++ src/main/java/com/JuanHDSM/Vendas/domain/dtos/RequestProductDTO.java
@@ -4,7 +4,7 @@ import com.JuanHDSM.Vendas.domain.entities.Product;
 
 import java.math.BigDecimal;
 
-public record RequestProductDTO(String description, BigDecimal price) {
+public record RequestProductDTO(String description, Double price) {
     public static RequestProductDTO fromRequestProductDTO(Product entity) {
         return new RequestProductDTO(
                 entity.getDescription(),
